Identity error school Description length

+++ src/Presentation/ViewModel/Identity/ProfileSettingsResponseViewModel.cs
@@ -5,6 +5,6 @@ namespace GamaEdtech.Presentation.ViewModel.Identity
         public int? CountryId { get; set; }
         public int? CityId { get; set; }
         public int? StateId { get; set; }
-        public int? SchoolId { get; set; }
+        public long? SchoolId { get; set; }
     }
 }
